+++ src/bible/db/models/verse.nim
@@ -5,33 +5,49 @@ import pkg/norm/[
 
 type
   Verse* = ref object of Model
-    ## A verse of a book in document
+    ## A chapter of a book in document
     docName*: string ## The name of the document
-    bookNumber*: int
+    bookShortName*: string
     chapter*: int
-    verse*: int
+    number*: int
     text*: string
 
 
 proc newVerse*(
-  docName: string;
-  bookNumber, chapter, verse: int;
+  docName, bookShortName: string;
+  chapter, number: int;
   text: string
 ): Verse =
   ## Creates new `Verse`
   new result
   result.docName = docName
-  result.bookNumber = bookNumber
+  result.bookShortName = bookShortName
   result.chapter = chapter
-  result.verse = verse
+  result.number = number
   result.text = text
 
 proc newVerse*: Verse =
   ## Creates new blank `Verse`
   newVerse(
     docName = "",
-    bookNumber = 0,
+    bookShortName = "",
     chapter = 0,
-    verse = 0,
+    number = 0,
     text = "",
   )
+
+import bible/db
+
+proc getAllBookVerses*(doc, bookShortName: string; chapter: int): seq[Verse] =
+  result = @[newVerse()]
+  try:
+    inDb: dbConn.select(
+      result,
+      "Verse.docName = ? and Verse.bookShortName = ? and Verse.chapter = ?",
+      dbValue doc, dbValue bookShortName, dbValue chapter
+    )
+  except: discard
+  if result[0].text.len == 0:
+    discard pop result
+
+# todo: change book id from booknum to book short name